General/multi/2/python/requirement.txt
backtrader
ccxt
pandas